Fix visit max from opt-prams gen

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/codegen/JvmMethodGen.java
@@ -2198,7 +2198,7 @@ public class JvmMethodGen {
         mv.visitVarInsn(ALOAD, 0);
 
         // load value[i]
-        mv.visitLdcInsn(paramIndex);
+        mv.visitLdcInsn((long) paramIndex);
         mv.visitInsn(L2I);
         mv.visitInsn(AALOAD);
     }
