+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/codegen/JvmValueGen.java
@@ -385,7 +385,7 @@ class JvmValueGen {
                     mv.visitVarInsn(ALOAD, 3);
 
                     // load j'th parameter
-                    mv.visitLdcInsn(j);
+                    mv.visitLdcInsn((long)j);
                     mv.visitInsn(L2I);
                     mv.visitInsn(AALOAD);
                     addUnboxInsn(mv, pType, useBString);
